Move map predicate and base it off the min/max
This predicate still wouldn't work all that well with a fully SeriesID sorted
block (so accress row groups). But let's fix that separately
